fix(api token): move angular token from shared lib to frontend-angular shared lib to avoid conflict of express and angular imports
