+++ libs/frontend-angular/shared/config/README.md
@@ -1,0 +1,3 @@
+# frontend-angular-shared-config
+
+This library was generated with [Nx](https://nx.dev).
