+++ libs/frontend-angular/shared/config/project.json
@@ -1,0 +1,20 @@
+{
+  "name": "frontend-angular-shared-config",
+  "$schema": "../../../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "library",
+  "sourceRoot": "libs/frontend-angular/shared/config/src",
+  "prefix": "carplates",
+  "targets": {
+    "lint": {
+      "executor": "@nrwl/linter:eslint",
+      "outputs": ["{options.outputFile}"],
+      "options": {
+        "lintFilePatterns": [
+          "libs/frontend-angular/shared/config/**/*.ts",
+          "libs/frontend-angular/shared/config/**/*.html"
+        ]
+      }
+    }
+  },
+  "tags": ["scope:frontend-angular type:constants"]
+}
