+++ libs/frontend-angular/shared/config/src/index.ts
@@ -1,0 +1,4 @@
+import { InjectionToken } from '@angular/core';
+export const BASE_API_TOKEN = new InjectionToken<string>('Base API');
+
+export const API_URL = 'http://localhost:8080/api';
