+++ libs/shared/common/constants/project.json
@@ -10,20 +10,6 @@
       "options": {
         "lintFilePatterns": ["libs/shared/common/constants/**/*.ts"]
       }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "libs/shared/common/constants/jest.config.ts",
-        "passWithNoTests": true
-      },
-      "configurations": {
-        "ci": {
-          "ci": true,
-          "codeCoverage": true
-        }
-      }
     }
   },
   "tags": ["scope:shared, type:constants"]
