+++ libs/shared/common/constants/tsconfig.json
@@ -5,9 +5,6 @@
   "references": [
     {
       "path": "./tsconfig.lib.json"
-    },
-    {
-      "path": "./tsconfig.spec.json"
     }
   ]
 }
