--- libs/shared/common/enums/jest.config.ts
@@ -1,11 +1,0 @@
-/* eslint-disable */
-export default {
-  displayName: 'shared-common-enums',
-  preset: '../../../../jest.preset.js',
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]sx?$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../../coverage/libs/shared/common/enums',
-};
