+++ libs/shared/common/enums/tsconfig.json
@@ -5,9 +5,6 @@
   "references": [
     {
       "path": "./tsconfig.lib.json"
-    },
-    {
-      "path": "./tsconfig.spec.json"
     }
   ]
 }
