+++ libs/shared/common/enums/tsconfig.lib.json
@@ -6,6 +6,6 @@
     "declaration": true,
     "types": ["node"]
   },
-  "exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"],
+  "exclude": [],
   "include": ["src/**/*.ts"]
 }
